requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.18.0
pydantic==2.5.3
sqlalchemy==2.0.25